Refactor: Move accordion components to own directories
This commit reorganizes the folder, kernel, and tag accordion components into their own dedicated directories for better code organization and maintainability.
